pyproject.toml
[project]
name = "actio-marketing-panel"
version = "1.0.0"
description = "Marketing Intelligence panel — GA4 / GSC / Google Ads → SQLite → daily LLM CMO report → email + Pushover + Chainlit dashboard"
authors = [{name = "Tom Lebioda", email = "hello@tomlebioda.com"}]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.2.0",
    "chainlit>=1.3",
    "plotly>=5.24",
    "pandas>=2.2",
    "google-analytics-data>=0.18",
    "google-api-python-client>=2.100",
    "google-ads>=25.0",
    "google-auth-oauthlib>=1.2",
    "python-dotenv>=1.0",
    "tabulate>=0.9",
    "httpx>=0.27",
    "markdown>=3.6",
]

[tool.uv]
package = false
